fix: resolve linting issues in marketplace filters - replace any type and fix useEffect dependencies
